handle O_APPEND better

+++ src/client/include/sdfs_fs.h
@@ -76,7 +76,7 @@ public:
 			cb_async_finished_t cb_finished, void* arg);
 
 	async_handle_t write(
-			unsigned long ino, size_t offset, size_t size, const char* buf,
+			unsigned long ino, off_t offset, size_t size, const char* buf,
 			cb_async_finished_t cb_finished, void* arg);
 
 	async_handle_t truncate(
